machine_learning/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
keras==3.8.0  # Keep Keras for now, even without TensorFlow
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-python-headless==4.11.0.86
opt-einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.38
sympy==1.13.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.75
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
flask-cors==3.0.10
psycopg2-binary==2.9.8
python-dotenv
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
waitress==2.1.2
gunicorn==21.2.0